refactor: ChatMessagesContainer 컴포넌트 리팩터링
