+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/_components/ChatFooter/ChatFooter.tsx
@@ -4,17 +4,20 @@ import MessageTextarea from '../MessageTextarea';
 import UtilsMenu from '../UtilsMenu';
 import ContextMenu from '../ContextMenu';
 import { useWorkspaceUserId } from '@/hooks/useWorkspaceUserId';
+import { useParams } from 'next/navigation';
 
 type ChatFooterProps = {
-  id: string;
   handleOpenUtil: () => void;
 };
 
-const ChatFooter = ({ id, handleOpenUtil }: ChatFooterProps) => {
+const ChatFooter = ({ handleOpenUtil }: ChatFooterProps) => {
+  const { id: channelId } = useParams();
+  const stringId = Array.isArray(channelId) ? channelId[0] : channelId;
+
   const workspaceUserId = useWorkspaceUserId();
   const ref = useRef<HTMLTextAreaElement>(null);
   const { mutate: mutateChatMessage } = useMutationChatMessage({
-    channel_id: Number(id),
+    channel_id: Number(stringId),
     workspace_user_id: workspaceUserId
   });
