+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/_components/ChatMessagesWrapper/ChatMessagesWrapper.tsx
@@ -1,7 +1,28 @@
 import { StrictPropsWithChildren } from '@/types/common';
-import { forwardRef } from 'react';
+import { forwardRef, useState } from 'react';
+import ChatFooter from '../ChatFooter';
 
-const ChatMessagesWrapper = forwardRef<HTMLDivElement, StrictPropsWithChildren>(({ children }, ref) => {
+export const ChatMessagesContainer = ({ children }: StrictPropsWithChildren) => {
+  const [isOpenUtil, setIsOpenUtil] = useState(false);
+
+  const handleOpenUtil = () => {
+    setIsOpenUtil((prev) => !prev);
+  };
+
+  return (
+    <div
+      className={`flex flex-col flex-grow h-[calc(100dvh+42px)] transform ease-in-out duration-300 ${
+        isOpenUtil ? 'translate-y-[-96px]' : 'translate-y-[0px]'
+      }`}
+    >
+      {children}
+      <ChatFooter handleOpenUtil={handleOpenUtil} />
+      {isOpenUtil && <div className="fixed top-0 left-0 w-full h-full z-40" onClick={handleOpenUtil} />}
+    </div>
+  );
+};
+
+export const ChatMessagesWrapper = forwardRef<HTMLDivElement, StrictPropsWithChildren>(({ children }, ref) => {
   return (
     <article className="flex-grow overflow-y-auto px-4">
       <div className="relative flex flex-col gap-6 py-4" ref={ref}>
@@ -10,5 +31,3 @@ const ChatMessagesWrapper = forwardRef<HTMLDivElement, StrictPropsWithChildren>(
     </article>
   );
 });
-
-export default ChatMessagesWrapper;
