+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/_components/ChatMessagesWrapper/index.ts
@@ -1,1 +1,1 @@
-export { default } from './ChatMessagesWrapper';
+export * from './ChatMessagesWrapper';
